+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.0
+* Add Android/iOS platform custom view support
+* Upgrade scankit android SDK to 2.1.0.300
+
+
 ## 1.1.0
 * Add Null Safety support
 * Upgrade the Android SDK version number of HUAWEI ScanKit to support full-screen scanning